# web-app/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "2.30.0",
    "title": "DigiGrotesk API",
    "description": "Utilize esta documentação para aulixio ao uso da API do sistema!"
  },
  "host": "localhost:3000",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "securityDefinitions": {},
  "consumes": [],
  "produces": [],
  "paths": {
...
    "/semester/{id}/subjects": {
      "get": {
        "summary" : "Retorna as disciplinas de um período específico",
        "description": "Retorna todas as disciplinas e suas informações do período especificado",
        "tags": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true
          },{
            "$ref": "#/parameters/x-key-api"
          },
          {
            "$ref": "#/parameters/x-token-api"
          }
        ],
        "responses": {
          "200": {
            "description": "Retorna um objeto subjects contendo um array com todos os disciplina do semestre selecionado",
            "schema": {
              "type": "object",
              "properties": {
                "subjects": {
                  "type": "array",
                  "example": [{}]
                }
              }
            }
          },
          "400": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "array",
                  "example": "É obrigatório passar o id do período na requisição para buscar as disciplinas!"
                }
              }
            }
          }
        }
      }
    },
...
    "/subject/": {
      "get": {
        "tags": [],
        "summary": "Retorna todas as disciplinas do usuário",
        "description": "Retorna todas as informações de todas as disciplinas do usuário",
        "parameters": [
          {
            "$ref": "#/parameters/x-key-api"
          },
          {
            "$ref": "#/parameters/x-token-api"
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "type": "object",
              "properties": {
                "subjects": {
                  "type": "array",
                  "example":[
                    {
                      "name": "Nome da disciplina",
                      "teacher": "Nome do professor",
                      "school": "Noma da instituição de ensino",
                      "author": "id do usuário",
                      "semester": "id do período",
                      "start": "2020-06-01",
                      "end": "2020-12-31",
                      "create": "2020-06-05",
                      "finalGrade": 7.8,
                      "grades": [
                        {
                          "value": 8.0,
                          "weigth": 5
                        }
                      ]
                    }
                  ]}
              }
            }
          },
          "500": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Erro interno do servidor"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [],
        "summary": "Criar disciplina",
        "description": "Cria uma disciplina para o usuário",
        "parameters": [
          {
            "$ref": "#/parameters/x-key-api"
          },
          {
            "$ref": "#/parameters/x-token-api"
          },
          {
            "in": "body",
            "name": "name",
            "required": true,
            "example": "Construção de Software",
            "description": "Nome da disciplina"
          },
          {
            "in": "body",
            "name": "semester",
            "required": true,
            "example": "abcd1234",
            "description": "Id do período em que a disciplina pertence"
          },
          {
            "in": "body",
            "name": "start",
            "required": true,
            "example": "2020-06-01",
            "description": "Data de início da disciplina"
          },
          {
            "in": "body",
            "name": "end",
            "required": true,
            "example": "2020-12-01",
            "description": "Data de término da disciplina"
          },
          {
            "in": "body",
            "name": "teacher",
            "required": false,
            "example": "Walisson",
            "description": "Nome do professor que vai ministrar a disciplina"
          },
          {
            "in": "body",
            "name": "school",
            "required": false,
            "example": "UFG - INF",
            "description": "Nome da instituição de ensino"
          }
        ],
        "responses": {
          "201": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Disciplina criada!"
                }
              }
            }
          },
          "500": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "É obrigatório passar uma disciplina para ser salva!"
                }
              }
            }
          }
        }
      }
    },
    "/subject/{id}": {
      "get": {
        "tags": [],
        "summary": "Retorna uma disciplina em específico",
        "description": "Retorna a disciplina correspondente ao id passado",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "$ref": "#/parameters/x-key-api"
          },
          {
            "$ref": "#/parameters/x-token-api"
          }
        ],
        "responses": {
          "201": {
            "schema": {
              "type": "object",
              "properties": {
                "subjects": {
                  "type": "array",
                  "example":
                    {
                      "name": "Nome da disciplina",
                      "teacher": "Nome do professor",
                      "school": "Noma da instituição de ensino",
                      "author": "id do usuário",
                      "semester": "id do período",
                      "start": "2020-06-01",
                      "end": "2020-12-31",
                      "create": "2020-06-05",
                      "finalGrade": 7.8,
                      "grades": [
                        {
                          "value": 8.0,
                          "weigth": 5
                        }
                      ]
                    }
                  }
              }
            }
          },
          "400": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "É obrigatório passar um id na requisição para buscar uma disciplina específica!"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [],
        "summary": "Atualizar disciplina",
        "description": "Permite atualizar nome, data de inicio e término, professor e instituição de ensino da disciplina",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "$ref": "#/parameters/x-key-api"
          },
          {
            "$ref": "#/parameters/x-token-api"
          },
          {
            "in": "body",
            "name": "name",
            "required": false,
            "example": "Construção de Software",
            "description": "Nome da disciplina"
          },
          {
            "in": "body",
            "name": "start",
            "required": false,
            "example": "2020-06-01",
            "description": "Data de início da disciplina"
          },
          {
            "in": "body",
            "name": "end",
            "required": false,
            "example": "2020-12-01",
            "description": "Data de término da disciplina"
          },
          {
            "in": "body",
            "name": "teacher",
            "required": false,
            "example": "Walisson",
            "description": "Nome do professor que vai ministrar a disciplina"
          },
          {
            "in": "body",
            "name": "school",
            "required": false,
            "example": "UFG - INF",
            "description": "Nome da instituição de ensino"
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Disciplina atualizada!"
                }
              }
            }
          },
          "400": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "É obrigatório passar um id na requisição para buscar uma disciplina específica!"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [],
        "summary": "Deletar disciplina",
        "description": "Deleta uma disciplina específica do usuário e todas as informações pertencentes a ela (tarefas, aulas, provas)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "$ref": "#/parameters/x-key-api"
          },
          {
            "$ref": "#/parameters/x-token-api"
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Disciplina exlcuida!"
                }
              }
            }
          }
        }
      }
    },
    "/subject/{id}/exams": {
      "get": {
        "tags": [],
        "summary": "Retorna todas as provas da disciplina",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "$ref": "#/parameters/x-key-api"
          },
          {
            "$ref": "#/parameters/x-token-api"
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "type": "object",
              "properties": {
                "exams": {
                  "type": "array",
                  "example":
                  [{
                    "name": "Nome de identificação da prova",
                    "subject": {
                      "id": "IDDISCIPLINA",
                      "author": "IDUSUARIO"
                    },
                    "theme": "Tema da prova",
                    "date": "2020-07-28",
                    "create": "2020-06-10"
                  }]
                }
              }
            }
          },
          "400": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "É obrigatório passar o id da disciplina na requisição para buscar as provas!"
                }
              }
            }
          }
        }
      }
    },
    "/subject/{id}/grade": {
      "put": {
        "tags": [],
        "summary": "Registrar notas da disciplina",
        "description": "Atualiza a nota da disciplina",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "$ref": "#/parameters/x-key-api"
          },
          {
            "$ref": "#/parameters/x-token-api"
          },
          {
            "in": "body",
            "name": "grades",
            "required": true,
            "example": [{
              "value": 8.2,
              "weigth": 5
            }],
            "description": "Array de objetos que tem o atributo value (valor da nota) e weigth (peso da nota)"
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Notas registradas!"
                }
              }
            }
          },
          "400": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "É obrigatório passar as notas para que sejam registradas na disciplina!"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [],
        "summary": "Retorna a nota da disciplina",
        "description": "Retorna a nota final da disciplina e as notas calculadas para gerar as notas finais",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "$ref": "#/parameters/x-key-api"
          },
          {
            "$ref": "#/parameters/x-token-api"
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "Nome da disciplina"
                },
                "finalGrade": {
                  "type": "string",
                  "example": 7.8
                },
                "grades": {
                  "type": "array",
                  "example": [{
                    "value": 8.0,
                    "weigth": 5
                  }]
                }
              }
            }
          },
          "400": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "É obrigatório passar o id da disciplina na requisição para buscar as notas!"
                }
              }
            }
          }
        }
      }
    },
...
    "/subject/{id}/absence": {
      "get": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/subject/{id}/absence/percent": {
      "get": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
...
  "parameters": {
    "x-key-api": {
      "name": "x-key-api",
      "in": "header",
      "description": "API Key do seu usuário",
      "required": true,
      "type": "string"
    },
    "x-token-api": {
      "name": "x-token-api",
      "in": "header",
      "description": "API Key do seu usuário",
      "required": true,
      "type": "string"
    }
  }
}